fix failing on error
Previously we piped to the command `cat > /dev/null` (or `cat -`). This meant the return value was always the return value of cat, hence everything was considered successful.

With this change, we just redirect to the file /dev/null or the file descriptor for stdout. Hence, if a function fails (returns a non-zero value) the script will actually terminate, as it should.